Lots of tweaking of design and animation
Added a wrapper for images that will break them out.
Switched out the secondary typeface Aleo for Zilla Slab. Updated the
homepage and quotation with this.
Added a bouncing animation for the homepage down arrow, made the
homepage "entries" slightly smaller so they take up around 88% of the
total height of the screen - displaying the next/previous entry for the
user. Adjusted the made in and version numbers to fit.
Reduced the width of content text, increased the font-size to make it
more readable.
Slightly increased the speed of the small-screen nav menu fly-in.
Centered all content headings and quotes.